Update the RPC ESModule so that it can follow the ERA easily
