+++ Geometry/RPCGeometryBuilder/plugins/RPCGeometryESModule.cc
@@ -8,20 +8,46 @@ DD4hep part added to the original old file (DD version) made by M. Maggi (INFN B
 //          Created:  Fri, 20 Sep 2019 
 //          Modified: Fri, 29 May 2020, following what Sunanda Banerjee made in PR #29842 PR #29943 and Ianna Osborne in PR #29954      
 */
-#include "Geometry/RPCGeometryBuilder/plugins/RPCGeometryESModule.h"
+#include "CondFormats/GeometryObjects/interface/RecoIdealGeometry.h"
+#include "DetectorDescription/Core/interface/DDCompactView.h"
+#include "DetectorDescription/DDCMS/interface/DDCompactView.h"
+#include "FWCore/Framework/interface/ESProducer.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/ESTransientHandle.h"
+#include "FWCore/Framework/interface/ESHandle.h"
+#include "FWCore/Framework/interface/ModuleFactory.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
+#include "Geometry/MuonNumbering/interface/MuonGeometryConstants.h"
+#include "Geometry/Records/interface/IdealGeometryRecord.h"
+#include "Geometry/Records/interface/MuonGeometryRecord.h"
+#include "Geometry/Records/interface/RPCRecoGeometryRcd.h"
 #include "Geometry/RPCGeometryBuilder/src/RPCGeometryBuilder.h"
 #include "Geometry/RPCGeometryBuilder/src/RPCGeometryBuilderFromCondDB.h"
-#include "Geometry/MuonNumbering/interface/MuonGeometryConstants.h"
-#include "Geometry/MuonNumbering/interface/MuonBaseNumber.h"
-#include <FWCore/Framework/interface/EventSetup.h>
-#include "FWCore/Framework/interface/ESTransientHandle.h"
-#include <FWCore/Framework/interface/ESHandle.h>
-#include <FWCore/Framework/interface/ModuleFactory.h>
-#include <FWCore/Framework/interface/ESProducer.h>
+#include "Geometry/RPCGeometry/interface/RPCGeometry.h"
 #include <unordered_map>
 #include <memory>
 
-using namespace edm;
+class RPCGeometryESModule : public edm::ESProducer {
+public:
+  RPCGeometryESModule(const edm::ParameterSet& p);
+
+  static void fillDescriptions(edm::ConfigurationDescriptions&);
+
+  std::unique_ptr<RPCGeometry> produce(const MuonGeometryRecord& record);
+
+private:
+  //DDD
+  edm::ESGetToken<DDCompactView, IdealGeometryRecord> idealGeomToken_;
+  edm::ESGetToken<MuonGeometryConstants, IdealGeometryRecord> dddConstantsToken_;
+  // dd4hep
+  edm::ESGetToken<cms::DDCompactView, IdealGeometryRecord> idealDD4hepGeomToken_;
+  // Reco
+  edm::ESGetToken<RecoIdealGeometry, RPCRecoGeometryRcd> recoIdealToken_;
+
+  const bool useDDD_;
+  const bool useDD4hep_;
+};
 
 RPCGeometryESModule::RPCGeometryESModule(const edm::ParameterSet& p)
     : useDDD_{p.getUntrackedParameter<bool>("useDDD", true)},
@@ -35,6 +61,13 @@ RPCGeometryESModule::RPCGeometryESModule(const edm::ParameterSet& p)
   } else {
     cc.setConsumes(recoIdealToken_);
   }
+}
+
+void RPCGeometryESModule::fillDescriptions(edm::ConfigurationDescriptions& descriptions) {
+  edm::ParameterSetDescription desc;
+  desc.addUntracked<bool>("useDDD", true);
+  desc.addUntracked<bool>("useDD4hep", false);
+  descriptions.add("RPCGeometryESModule", desc);
 }
 
 std::unique_ptr<RPCGeometry> RPCGeometryESModule::produce(const MuonGeometryRecord& record) {
