+++ Geometry/RPCGeometryBuilder/python/rpcGeometry_cfi.py
@@ -1,10 +1,5 @@
-import FWCore.ParameterSet.Config as cms
+from Geometry.RPCGeometryBuilder.RPCGeometryESModule_cfi import RPCGeometryESModule
 
-#
-# This cfi should be included to build the RPC geometry model.
-#
-RPCGeometryESModule = cms.ESProducer("RPCGeometryESModule",
-    useDDD = cms.untracked.bool(True)
-)
+from Configuration.ProcessModifiers.dd4hep_cff import dd4hep
 
-
+dd4hep.toModify(RPCGeometryESModule, useDDD = False, useDD4hep = True)
