+++ Geometry/RPCGeometryBuilder/test/python/validateRPCGeometryDDD_cfg.py
@@ -1,0 +1,23 @@
+import FWCore.ParameterSet.Config as cms
+from Configuration.Eras.Era_Run3_cff import Run3
+
+process = cms.Process('VALID',Run3)
+
+process.source = cms.Source('EmptySource')
+
+process.maxEvents = cms.untracked.PSet(
+    input = cms.untracked.int32(1)
+    )
+
+process.load('Configuration.Geometry.GeometryExtended2021_cff')
+process.load("FWCore.MessageLogger.MessageLogger_cfi")
+process.load("Geometry.MuonNumbering.muonGeometryConstants_cff")
+process.load("Geometry.RPCGeometryBuilder.rpcGeometry_cfi")
+
+process.valid = cms.EDAnalyzer("RPCGeometryValidate",
+                               infileName = cms.untracked.string('cmsRecoGeom-2021.root'),
+                               outfileName = cms.untracked.string('validateRPCGeometry.root'),
+                               tolerance = cms.untracked.int32(7)
+                               )
+
+process.p = cms.Path(process.valid)
