--- Geometry/RPCGeometryBuilder/test/python/validateRPCGeometry_cfg.py
@@ -1,47 +1,0 @@
-import FWCore.ParameterSet.Config as cms
-
-process = cms.Process('VALID')
-
-process.source = cms.Source('EmptySource')
-
-process.maxEvents = cms.untracked.PSet(
-    input = cms.untracked.int32(1)
-    )
-
-process.load('Configuration.StandardSequences.DD4hep_GeometrySim_cff')
-process.load("FWCore.MessageLogger.MessageLogger_cfi")
-process.load("Geometry.MuonNumbering.muonGeometryConstants_cff")
-
-process.RPCGeometryESProducer = cms.ESProducer("RPCGeometryESModule",
-                                               DDDetector = cms.ESInputTag('',''),
-                                               attribute = cms.string('ReadOutName'),
-                                               value = cms.string('MuonRPCHits'),
-                                               useDDD = cms.untracked.bool(False),
-                                               useDD4hep = cms.untracked.bool(True)
-                                              )
-
-process.DDCompactViewESProducer = cms.ESProducer("DDCompactViewESProducer",
-                                                 appendToDataLabel = cms.string('')
-)
-
-process.DDSpecParRegistryESProducer = cms.ESProducer("DDSpecParRegistryESProducer",
-                                                     appendToDataLabel = cms.string('')
-                                                     )
-
-process.muonGeometryConstants.fromDD4Hep = True
-
-process.test = cms.EDAnalyzer("DDTestMuonNumbering")
-
-#
-# Note: Please, download the geometry file from a location
-#       specified by Fireworks/Geometry/data/download.url
-#
-# For example: cmsRun $CMSSW_RELEASE_BASE/src/Fireworks/Geometry/python/dumpRecoGeometry_cfg.py tag=2021
-#
-process.valid = cms.EDAnalyzer("RPCGeometryValidate",
-                               infileName = cms.untracked.string('cmsRecoGeom-2021.root'),
-                               outfileName = cms.untracked.string('validateRPCGeometry.root'),
-                               tolerance = cms.untracked.int32(7)
-                               )
-
-process.p = cms.Path(process.valid)
